// src/app/models/fornecedor.model.ts
export interface FornecedorModel {
  id: number;
  nome: string;
  cnpj: string;
  telefone: string;
  descricao: string;
  codigoAnel?: string;
  codigoBracelete?: string;
  codigoColar?: string;
  codigoBrinco?: string;
  codigoPulseira?: string;
  codigoPingente?: string;
  codigoConjunto?: string;
  codigoBerloque?: string;
  codigoPiercing?: string;
}